requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
motor==3.4.0
pymongo==4.6.3
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-dotenv==1.0.0
